libobs: Fix a warning

+++ libobs/obs-video.c
@@ -757,7 +757,7 @@ static inline void video_sleep(struct obs_core_video *video, bool raw_active,
 		int64_t diff;
 		memcpy(&diff, &udiff, sizeof(diff));
 		const uint64_t clamped_diff =
-			(diff > (int64_t)interval_ns) ? diff : interval_ns;
+			(diff > (int64_t)interval_ns) ? (uint64_t)diff : interval_ns;
 		count = (int)(clamped_diff / interval_ns);
 		*p_time = cur_time + interval_ns * count;
 		render_too_slow ++;
